[refact] girafe.js: Moved inner code to srcsjs/factory.js
